+++ lib/flutter_smooth_drawer.dart
@@ -1,3 +1,4 @@
 
+export 'src/smooth_drawer_scope.dart';
 export 'src/smooth_hidden_drawer.dart';
 export 'src/smooth_hidden_drawer_controller.dart';